Rework the flask app to return an app. First iteration
