+++ airflow/__init__.py
@@ -10,28 +10,41 @@ __version__ = "1.5.2"
 import logging
 import os
 import sys
-from airflow.configuration import conf
+
+from airflow.configuration import conf, AirflowConfigException
 from airflow.models import DAG
 from flask.ext.admin import BaseView
-
+from importlib import import_module
+from utils import AirflowException
 
 DAGS_FOLDER = os.path.expanduser(conf.get('core', 'DAGS_FOLDER'))
 if DAGS_FOLDER not in sys.path:
     sys.path.append(DAGS_FOLDER)
 
-from airflow import default_login as login
-if conf.getboolean('webserver', 'AUTHENTICATE'):
-    try:
-        # Environment specific login
-        import airflow_login as login
-    except ImportError:
-        logging.error(
-            "authenticate is set to True in airflow.cfg, "
-            "but airflow_login failed to import")
+auth_backend = 'airflow.default_login'
+try:
+    auth_backend = conf.get('webserver', 'auth_backend')
+except AirflowConfigException:
+    if conf.getboolean('webserver', 'AUTHENTICATE'):
+        logging.warning("auth_backend not found in webserver config reverting to *deprecated*"
+                        " behavior of importing airflow_login")
+        auth_backend = "airflow_login"
+
+try:
+    login = import_module(auth_backend)
+except ImportError:
+    logging.critical(
+        "Cannot import authentication module %s. "
+        "Please correct your authentication backend or disable authentication",
+        auth_backend
+    )
+    if conf.getboolean('webserver', 'AUTHENTICATE'):
+        raise AirflowException("Failed to import authentication backend")
 
 
 class AirflowViewPlugin(BaseView):
     pass
+
 
 class AirflowMacroPlugin(object):
     def __init__(self, namespace):
